Implemented a refresh token
After user logs in the system now generates an additional JWT,
the refresh token. The token is placed in an Http-Only cookie for
the /refresh-token route.

On the client side, on every request that requires a valid
access token, the current access token is checked to see if
it has expired. If the token is no longer valid, the system
fetches a new one at /refresh-token before proceeding with
the original request.